% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_classes.R
\docType{class}
\name{Dataset-class}
\alias{Dataset-class}
\title{Dataset Class}
\value{

}
\description{
spatial-(temporal) data stored in NetCDF file. Assumes data is spatial (2D with coordinates) or spatial-temporal (+3D with coordinates and time dimension).
}
\section{Slots}{

\describe{
\item{\code{from}}{character.}

\item{\code{var}}{character.}

\item{\code{aggregation}}{list.}

\item{\code{filter}}{list.}

\item{\code{start}}{POSIXct.}

\item{\code{end}}{POSIXct.}
}}

