% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_classes.R
\docType{class}
\name{Frequency-class}
\alias{Frequency-class}
\title{Frequency}
\value{

}
\description{
Frequency
}
\section{Slots}{

\describe{
\item{\code{value}}{numeric.}

\item{\code{period}}{character.}
}}

