% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf_write_data.R
\name{ncdf.create}
\alias{ncdf.create}
\title{create ncdf}
\usage{
ncdf.create(file, domain_file)
}
\value{

}
\description{
Creates new NetCDF file from VIC domain file. Writes mask, grid, lat and lon variables.
}
