% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_tool}
\alias{run_tool}
\title{Run Tool}
\usage{
run_tool(config)
}
\arguments{
\item{config}{}
}
\value{

}
\description{
Run Tool
}
\examples{
# example of a validation
con <- system.file("extdata","validation_config.yml", package="VICvalicaliR")
run_tool(con)
con <- system.file("extdata","calibration_config.yml", package="VICvalicaliR")
files_to_copy <- c(system.file("extdata","vic_image.exe", package="VICvalicaliR"),system.file("extdata","vic_global_params.txt", package="VICvalicaliR"),
system.file("extdata","pr_WFDE5_3h_1993.nc", package="VICvalicaliR"),system.file("extdata","psurf_WFDE5_3h_1993.nc", package="VICvalicaliR"),
system.file("extdata","tas_WFDE5_3h_1993.nc", package="VICvalicaliR"),system.file("extdata","wind_WFDE5_3h_1993.nc", package="VICvalicaliR"),
system.file("extdata","lwdown_WFDE5_3h_1993.nc", package="VICvalicaliR"), system.file("extdata","swdown_WFDE5_3h_1993.nc", package="VICvalicaliR"),
system.file("extdata","vp_WFDE5_3h_1993.nc", package="VICvalicaliR"),
system.file("extdata","rhine_domain.nc", package="VICvalicaliR"),system.file("extdata","rhine_vic_params.nc", package="VICvalicaliR"),
system.file("extdata","rhine_lisflood_routing_params.nc", package="VICvalicaliR"),system.file("extdata","init_state.nc", package="VICvalicaliR") )
dir.create("./example_calibration", showWarnings=F)
for(fpath in files_to_copy){
  if(!file.exists(fpath)){
     system(sprintf("cp \%s ./example_calibration", fpath))
  }

}
run_tool(con)
}
