% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_classes.R
\docType{class}
\name{Validation-class}
\alias{Validation-class}
\title{Validation}
\value{

}
\description{
Validation
}
\section{Slots}{

\describe{
\item{\code{start}}{POSIXct.}

\item{\code{end}}{POSIXct.}
}}

