% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_tool}
\alias{run_tool}
\title{Run Tool}
\usage{
run_tool(config)
}
\arguments{
\item{config}{}
}
\value{

}
\description{
Run Tool
}
\examples{
# example of a validation
con <- system.file("extdata","validation_config.yml", package="VICvalicaliR")
run_tool(con)
}
